+++ features/mypage/mutations.ts
@@ -10,8 +10,13 @@ import {
 	uploadProfileImage,
 } from "./apis";
 import { useToast } from "@/providers/toast-provider";
+import { useUserStore } from "@/store/user.store";
 import { mypageQueryKeys } from "./queries";
 import { meetupDetailQueryKeys } from "../meetupDetail/queries";
+
+interface UsePatchUsersMeOptions {
+	onSuccessBeforeSync?: () => void;
+}
 
 export function useUploadProfileImage() {
 	const { handleShowToast } = useToast();
@@ -34,17 +39,22 @@ export function useUploadProfileImage() {
 	});
 }
 
-export function usePatchUsersMe() {
+export function usePatchUsersMe(options?: UsePatchUsersMeOptions) {
+	const queryClient = useQueryClient();
+	const setUser = useUserStore((state) => state.setUser);
 	const { handleShowToast } = useToast();
 
 	return useMutation({
 		mutationFn: patchUsersMe,
 
-		onSuccess: () => {
+		onSuccess: (updatedUser) => {
+			options?.onSuccessBeforeSync?.();
+			setUser(updatedUser);
 			handleShowToast({
 				message: "프로필이 수정되었습니다.",
 				status: "success",
 			});
+			queryClient.invalidateQueries({ queryKey: ["me"] });
 		},
 
 		onError: () => {
